pyproject.toml
[project]
name = "rat-bench-submission"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-core>=2.25.1",
    "azure-ai-textanalytics>=5.3.0",
    "google-genai>=1.31.0",
    "levenshtein>=0.27.1",
    "matplotlib>=3.10.5",
    "nltk>=3.9.1",
    "pandas>=2.3.2",
    "presidio-analyzer[gliner]>=2.2.359",
    "presidio-anonymizer>=2.2.359",
    "seaborn>=0.13.2",
    "vllm>=0.10.1.1",
    "transformers>=4.30.1",
    "fire>=0.7.1",
    "tokenizers>=0.13.1",
    "gradio>=5.44.1",
    "rouge-score>=0.1.2",
    "flair>=0.15.1",
    "llama-stack>=0.2.20",
    "huggingface-hub[cli]>=0.34.4",
    "ipywidgets>=8.1.7",
    "juliacall==0.9.30",
    "juliapkg==0.1.22",
    "scrubadub>=2.0.1",
    "click>=8.2.1",
    "datasets>=4.1.1",
    "gensim>=4.4.0",
    "pandarallel>=1.6.5",
    "sacremoses>=0.1.1",
    "correctmatch>=1.4.1",
    "ipykernel>=6.30.1",
]
